Eliminando dependencias de votainteligente y pasandolo a un votai_utils
