+++ votainteligente/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import
 
-from .celery import app as celery_app
+from votai_utils.celery import app as celery_app
